+++ extensions/youtube/src/main/java/app/revanced/extension/youtube/patches/components/LayoutComponentsFilter.java
@@ -41,6 +41,8 @@ public final class LayoutComponentsFilter extends Filter {
     private final StringFilterGroup horizontalShelves;
     private final ByteArrayFilterGroup ticketShelf;
     private final StringFilterGroup chipBar;
+    private final StringFilterGroup channelProfile;
+    private final ByteArrayFilterGroupList channelProfileBuffer;
 
     public LayoutComponentsFilter() {
         exceptions.addPatterns(
@@ -82,18 +84,13 @@ public final class LayoutComponentsFilter extends Filter {
                 "poll_post_responsive_root.eml"
         );
 
-        final var communityGuidelines = new StringFilterGroup(
-                Settings.HIDE_COMMUNITY_GUIDELINES,
-                "community_guidelines"
-        );
-
         final var subscribersCommunityGuidelines = new StringFilterGroup(
                 Settings.HIDE_SUBSCRIBERS_COMMUNITY_GUIDELINES,
                 "sponsorships_comments_upsell"
         );
 
-        final var channelMemberShelf = new StringFilterGroup(
-                Settings.HIDE_CHANNEL_MEMBER_SHELF,
+        final var channelMembersShelf = new StringFilterGroup(
+                Settings.HIDE_MEMBERS_SHELF,
                 "member_recognition_shelf"
         );
 
@@ -140,13 +137,13 @@ public final class LayoutComponentsFilter extends Filter {
         );
 
         final var latestPosts = new StringFilterGroup(
-                Settings.HIDE_HIDE_LATEST_POSTS,
+                Settings.HIDE_LATEST_POSTS,
                 "post_shelf"
         );
 
-        final var channelGuidelines = new StringFilterGroup(
-                Settings.HIDE_HIDE_CHANNEL_GUIDELINES,
-                "channel_guidelines_entry_banner"
+        final var channelLinksPreview = new StringFilterGroup(
+                Settings.HIDE_LINKS_PREVIEW,
+                "attribution.eml"
         );
 
         final var emergencyBox = new StringFilterGroup(
@@ -201,7 +198,6 @@ public final class LayoutComponentsFilter extends Filter {
                 "image_shelf"
         );
 
-
         final var timedReactions = new StringFilterGroup(
                 Settings.HIDE_TIMED_REACTIONS,
                 "emoji_control_panel",
@@ -228,7 +224,6 @@ public final class LayoutComponentsFilter extends Filter {
                 "sponsorships"
         );
 
-
         final var channelWatermark = new StringFilterGroup(
                 Settings.HIDE_VIDEO_CHANNEL_WATERMARK,
                 "featured_channel_watermark_overlay"
@@ -242,6 +237,22 @@ public final class LayoutComponentsFilter extends Filter {
         final var searchResultRecommendationLabels = new StringFilterGroup(
                 Settings.HIDE_SEARCH_RESULT_RECOMMENDATION_LABELS,
                 "endorsement_header_footer.eml"
+        );
+
+        channelProfile = new StringFilterGroup(
+                null,
+                "channel_profile.eml",
+                "page_header.eml"
+        );
+        channelProfileBuffer = new ByteArrayFilterGroupList();
+        channelProfileBuffer.addAll(new ByteArrayFilterGroup(
+                        Settings.HIDE_VISIT_STORE_BUTTON,
+                        "header_store_button"
+                ),
+                new ByteArrayFilterGroup(
+                        Settings.HIDE_VISIT_COMMUNITY_BUTTON,
+                        "community_button"
+                )
         );
 
         horizontalShelves = new StringFilterGroup(
@@ -267,7 +278,6 @@ public final class LayoutComponentsFilter extends Filter {
                 searchResultRecommendationLabels,
                 latestPosts,
                 channelWatermark,
-                communityGuidelines,
                 playables,
                 quickActions,
                 relatedVideos,
@@ -280,12 +290,13 @@ public final class LayoutComponentsFilter extends Filter {
                 subscribersCommunityGuidelines,
                 subscriptionsChipBar,
                 chipBar,
-                channelGuidelines,
+                channelLinksPreview,
+                channelProfile,
                 audioTrackButton,
                 artistCard,
                 timedReactions,
                 imageShelf,
-                channelMemberShelf,
+                channelMembersShelf,
                 forYouShelf,
                 horizontalShelves
         );
@@ -307,6 +318,10 @@ public final class LayoutComponentsFilter extends Filter {
         // Filter them separately here.
         if (matchedGroup == notifyMe || matchedGroup == inFeedSurvey || matchedGroup == expandableMetadata) {
             return true;
+        }
+
+        if (matchedGroup == channelProfile) {
+            return channelProfileBuffer.check(protobufBufferArray).isFiltered();
         }
 
         if (exceptions.matches(path)) return false; // Exceptions are not filtered.
@@ -333,7 +348,7 @@ public final class LayoutComponentsFilter extends Filter {
      * Injection point.
      * Called from a different place then the other filters.
      */
-    public static boolean filterMixPlaylists(final Object conversionContext, @Nullable final byte[] bytes) {
+    public static boolean filterMixPlaylists(Object conversionContext, @Nullable final byte[] bytes) {
         try {
             if (!Settings.HIDE_MIX_PLAYLISTS.get()) {
                 return false;